+++ examples/data-sources/pcd_networking_qos_policy/data-source.tf
@@ -1,0 +1,3 @@
+data "pcd_networking_qos_policy" "example" {
+  name = "tf-example-qos"
+}