+++ examples/resources/pcd_networking_qos_policy/resource.tf
@@ -1,0 +1,6 @@
+resource "pcd_networking_qos_policy" "example" {
+  name        = "tf-example-qos"
+  description = "QoS policy managed by Terraform"
+  shared      = true
+  tags        = ["terraform", "qos"]
+}